Move welcome Dialog into Room component
